+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## vNext
+## v0.6.0-rc.1 (2017-07-11) **[pre-release @next]**
 
 * **BREAKING** Upgrade dependencies for compatibility with Node 8 and TypeScript 2.4
     * Webpack 3